+++ test/pbkdf2_test.exs
@@ -2,6 +2,7 @@ defmodule Pbkdf2Test do
   use ExUnit.Case
   doctest Pbkdf2
 
+  import ExUnit.CaptureIO
   import Comeonin.BehaviourTestHelper
 
   test "implementation of Comeonin.PasswordHash behaviour" do
@@ -32,6 +33,17 @@ defmodule Pbkdf2Test do
     assert byte_size(Pbkdf2.gen_salt()) == 16
     assert byte_size(Pbkdf2.gen_salt(32)) == 32
     assert byte_size(Pbkdf2.gen_salt(64)) == 64
+  end
+
+  test "gen_salt prints warnings for salts that are too short" do
+    assert capture_io(:stderr, fn -> Pbkdf2.gen_salt(7) end) =~
+             "Using a salt less than 8 bytes long is not recommended"
+  end
+
+  test "gen_salt raises if salt is too long" do
+    assert_raise ArgumentError, fn ->
+      Pbkdf2.gen_salt(1025)
+    end
   end
 
   test "hashes with different lengths are correctly verified" do